message-processor/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pika==1.2.0
requests==2.26.0
PyJWT==2.1.0
Authlib==0.15.4
gunicorn==20.1.0
werkzeug==2.0.3